📦 cry: Generated migration to add the `create_at` and `admission_date` fields to the model (Person)
A Django migration was generated to add the `create_at` and `admission_date` fields to the model (Person). These fields are important for recording the creation date of a model instance and the admission date, respectively. This migration is part of evolving the database schema to accommodate new requirements and functionality.